Improve test script for standalone binary

+++ tests/test.sh
@@ -1,9 +1,19 @@
 #!/bin/bash
 
+IESCRYPT="iescrypt"
 FAILURE=0
+CURRDIR=${PWD}
+WORKDIR=$(mktemp -d)
+
+cd "${WORKDIR}"
+if test ! -d "${WORKDIR}";
+then
+    echo "failed to create work directory"
+    exit 1
+fi
 
 rm -f ekey ekey.pub
-if iescrypt gen-enc ekey && test -f ekey && test -f ekey.pub;
+if ${IESCRYPT} gen-enc ekey && test -f ekey && test -f ekey.pub;
 then
     echo "gen-enc test succeeded"
 else
@@ -12,7 +22,7 @@ else
 fi
 
 rm -f skey skey.pub
-if iescrypt gen-sig skey && test -f skey && test -f skey.pub;
+if ${IESCRYPT} gen-sig skey && test -f skey && test -f skey.pub;
 then
     echo "gen-enc test succeeded"
 else
@@ -25,7 +35,7 @@ echo "aZerty6" > passphrase
 dd if=/dev/urandom bs=1M count=100 > m.dat 2> /dev/null
 
 rm -f c.dat d.dat
-if iescrypt penc m.dat c.dat passphrase && test -f c.dat && iescrypt pdec c.dat d.dat passphrase && test -f d.dat && diff -q m.dat d.dat;
+if ${IESCRYPT} penc m.dat c.dat passphrase && test -f c.dat && ${IESCRYPT} pdec c.dat d.dat passphrase && test -f d.dat && diff -q m.dat d.dat;
 then
     echo "penc/pdec test succeeded"
 else
@@ -34,7 +44,7 @@ else
 fi
 
 rm -f c.dat d.dat
-if iescrypt enc m.dat c.dat ekey.pub && test -f c.dat && iescrypt dec c.dat d.dat ekey && test -f d.dat && diff -q m.dat d.dat;
+if ${IESCRYPT} enc m.dat c.dat ekey.pub && test -f c.dat && ${IESCRYPT} dec c.dat d.dat ekey && test -f d.dat && diff -q m.dat d.dat;
 then
     echo "enc/dec test succeeded"
 else
@@ -43,7 +53,7 @@ else
 fi
 
 rm -f m.dat.sig
-if iescrypt sig m.dat m.dat.sig skey && test -f m.dat.sig && test -n "$(iescrypt ver m.dat m.dat.sig skey.pub | grep 'Valid signature')";
+if ${IESCRYPT} sig m.dat m.dat.sig skey && test -f m.dat.sig && test -n "$(${IESCRYPT} ver m.dat m.dat.sig skey.pub | grep 'Valid signature')";
 then
     echo "sig/ver test succeeded"
 else
@@ -52,7 +62,7 @@ else
 fi
 
 rm -f c.dat d.dat m.dat.sig
-if iescrypt sig-penc m.dat c.dat skey passphrase && test -f c.dat && test -n "$(iescrypt pdec-ver c.dat d.dat passphrase skey.pub | grep 'Valid signature')" && diff -q m.dat d.dat;
+if ${IESCRYPT} sig-penc m.dat c.dat skey passphrase && test -f c.dat && test -n "$(${IESCRYPT} pdec-ver c.dat d.dat passphrase skey.pub | grep 'Valid signature')" && diff -q m.dat d.dat;
 then
     echo "sig-penc/pdec-ver test succeeded"
 else
@@ -61,7 +71,7 @@ else
 fi
 
 rm -f c.dat d.dat m.dat.sig
-if iescrypt sig-enc m.dat c.dat skey ekey.pub && test -f c.dat && test -n "$(iescrypt dec-ver c.dat d.dat ekey skey.pub | grep 'Valid signature')" && diff -q m.dat d.dat;
+if ${IESCRYPT} sig-enc m.dat c.dat skey ekey.pub && test -f c.dat && test -n "$(${IESCRYPT} dec-ver c.dat d.dat ekey skey.pub | grep 'Valid signature')" && diff -q m.dat d.dat;
 then
     echo "sig-enc/dec-ver test succeeded"
 else
@@ -70,10 +80,12 @@ else
 fi
 
 echo
+cd "${PWD}"
 if test ${FAILURE} -eq 1;
 then
     echo "FAILURE"
+    exit 1
 else
-    rm -f ekey ekey.pub skey skey.pub passphrase m.dat c.dat d.dat m.dat.sig
+    rm -fr "${WORKDIR}"
     echo "OK"
 fi
